feat(preventive-works): рефакторинг кода Обработчики событий вынесены в отдельные файлы
